Custom View for Human Hand
Changed the view for the human's had to a custom view so each card has an index that is the same as the index of the card in the array.

The user can now select a card from it's hand and can trail that card.

The "Computer Move" is setup to just trail the first card in the it's hand